requirements.txt
Flask==2.2.5
requests==2.32.3
PyMySql==1.1.1
cryptography==44.0.1
python-dotenv==1.0.1